3 tier architecture
