+++ backend/nokia-innovative-project/src/main/java/com/nokia/library/nokiainnovativeproject/services/BookService.java
@@ -1,0 +1,62 @@
+package com.nokia.library.nokiainnovativeproject.services;
+
+import com.nokia.library.nokiainnovativeproject.entities.Book;
+import com.nokia.library.nokiainnovativeproject.exceptions.ResourceNotFoundException;
+import com.nokia.library.nokiainnovativeproject.repositories.BookRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+
+import javax.validation.Valid;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+@Service
+@RequiredArgsConstructor
+public class BookService {
+
+	private final BookRepository bookRepository;
+
+	public List<Book> getAllBooks() {
+		return bookRepository.findAll();
+	}
+
+
+	public Book getBookById(@PathVariable Long id) {
+		return bookRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Book", "id", id));
+	}
+
+
+	public Book createBook(@Valid @RequestBody Book book) {
+		return bookRepository.save(book);
+	}
+
+
+	public ResponseEntity<Book> updateBook(
+			@PathVariable Long id,
+			@Valid @RequestBody Book bookDetails) throws ResourceNotFoundException {
+		Book book = bookRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Book", "id", id));
+
+		book.setTitle(bookDetails.getTitle());
+		book.setAuthorName(bookDetails.getAuthorName());
+		book.setAuthorSurname(bookDetails.getAuthorSurname());
+		final Book updatedBook = bookRepository.save(book);
+		return ResponseEntity.ok(updatedBook);
+	}
+
+	public Map<String, Boolean> deleteBook(
+			@PathVariable Long id) throws ResourceNotFoundException {
+		Book book = bookRepository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundException("Book", "id", id));
+
+		bookRepository.delete(book);
+		return new HashMap<String, Boolean>() {{
+			put("deleted", true);
+		}};
+	}
+}
